Updated structure and added databases(NOT WORK Properly now)
Needs to end
