+++ README.md
@@ -1,2 +1,4 @@
-Exercise 2 from programming part 2
+
+# Exercise 2 from programming part 2
+
 LPNU
