+++ requirements.txt
@@ -1,4 +1,5 @@
 
 # Usage: pip install -r requirements.txt
 
-Flask==1.0.2
+Flask==2.0
+Flask-SQLAlchemy==2.5.1
