--- rest-api/fish_shop/main.py
@@ -1,9 +1,0 @@
-from flask import Flask
-
-app = Flask(__name__)
-
-from controller import *
-
-if __name__ == "__main__":
-    app.run()
-
